pyproject.toml
[project]
name = "terminusgps-install"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.1",
    "pillow>=11.2.1",
    "python-terminusgps>=36.6.0",
]

[dependency-groups]
dev = [
    "cookiecutter>=2.6.0",
    "django-tailwind[reload]>=4.0.1",
]
